requirements.txt
django
djangorestframework
djangorestframework-simplejwt
python-dotenv
django-filter
drf-yasg